Virtual memory is working.
A few changes are needed but we're getting there:

Currently the processes have access to some working
memory they shouldn't have. Shouldn't be a hard fix.

And there is no way for them to create new l1 tables
for themself or another process.

Not yet a way to create a process though, so that's not
too bad.