Bunch of shit. Working.
There were/are problems with unmapping and I need a way to ensure
that the write buffer (?) has cleared. Currently yielding after
unmapping memory.

There are problems with the pool I think which I have temporaraly
solved by increasing the initial size so it doesn't need to grow.

Probably other things.

I added a way to let proc0 print debug messages through the kernel
with a syscall. It's not pretty.